requirements.txt
dnspython==2.2.1
mysql-connector-python==8.0.31
protobuf==3.20.1
pymongo==4.3.3
